% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exblockGEE.R
\name{exblockGEE_test}
\alias{exblockGEE_test}
\title{Tests for general hypothesis in repeated measures generalized linear models
with block invariant covariates}
\usage{
exblockGEE_test(
  data,
  between = NULL,
  within,
  dependent,
  link,
  Cont,
  r = rep(0, nrow(Cont)),
  eps = 1e-12,
  ex_col,
  ex_min,
  ex_max,
  var_ex = 0,
  test,
  grid = 100
)
}
\arguments{
\item{data}{Matrix object containing all variables.}

\item{between}{Integer vector containing position of between factors in data (column number).}

\item{within}{Integer vector containing position of within factors in data (column number).}

\item{dependent}{Integer defining the position of the dependent variable in data (column number).}

\item{link}{Character defining the link function, "pois" for Poisson, "ordinal" for cumulative logit.}

\item{Cont}{Numeric matrix defining contrasts R for the nullhypothesis Rb=r.}

\item{r}{Numeric vector r for the nullhypothesis Rb=r.}

\item{eps}{Numeric specifying the precision at which numeric algorithms stop.}

\item{ex_col}{Integer vector indicating the group means meant by external interval.}

\item{ex_min}{Numeric vector containing the lower bounds of external interval.}

\item{ex_max}{Numeric vector containing the upper bounds of external interval.}

\item{var_ex}{Numeric matrix equal to external variance.}

\item{test}{Character specifying which test to use, "Wald" for Wald test, "LM" for Lagrange-Multiplier test, "DRU" for difference test.}

\item{grid}{Number of grid points.}
}
\value{
Vector containing minimal test statistic, degrees of freedom, p value and convergence flag, 1 = converged, 0 = stopped after 10000 iterations.
}
\description{
Tests for general hypothesis in repeated measures generalized linear models
with block invariant covariates
}
\examples{

data<-cbind(rep(0:1,times=32),rep(0:1,each=2,times=16),rpois(64,lambda=3))

#overall test whether all coefficients are different from zero:
exblockGEE_test(data,within=c(1,2), dependent=c(3),link="pois",Cont=cbind(0,diag(3)),
ex_col = c(1),ex_min=2.7,ex_max=3.5,var_ex=1/1000,test="Wald")

exblockGEE_test(data,within=c(1,2), dependent=c(3),link="pois",Cont=cbind(0,diag(3)),
ex_col = c(1),ex_min=2.7,ex_max=3.5,var_ex=1/1000,test="LM")

exblockGEE_test(data,within=c(1,2), dependent=c(3),link="pois",Cont=cbind(0,diag(3)),
ex_col = c(1),ex_min=2.7,ex_max=3.5,var_ex=1/1000,test="DRU")
}
