% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exlm.R
\name{exlm}
\alias{exlm}
\title{Externally informed linear model}
\usage{
exlm(
  X,
  y,
  mom,
  I_min,
  I_max = I_min,
  grid_points = 10001,
  intercept = TRUE,
  var.e
)
}
\arguments{
\item{X}{Numeric matrix containing independent variables in colums.}

\item{y}{Numeric vector containing dependent variable.}

\item{mom}{Character vector of moments, "EY" = expected value of y, "EXi" = expected value of ith column of X,
"EXiY" = mixed moment of y and x_i, "EY2" = second moment of y, "VY" = variance of y,
"CXiY" = covariance of y and x_i, "RXiY" = correlation of y and x_i, "BXiY" = simple linear regression slope
of y on x_i. The index i must be replaced by a column number of X, e.g. "EX1Y" takes the first column of X.}

\item{I_min}{Numeric vector containing the lower bounds of external interval.}

\item{I_max}{Numeric vector containing the upper bounds of external interval.}

\item{grid_points}{Number of Grid points.}

\item{intercept}{Logical, indicates wheter intercept is used in model or not.}

\item{var.e}{Numeric matrix equal to external variance.}
}
\value{
Numeric matrix with lower and upper bounds of estimates, variances and confidence unions.
}
\description{
Computes intervals for estimates and their variances as well as confidence unions.
}
\examples{
x<-rnorm(100);y<-rnorm(100)
exlm(x,y,c("EY"),c(-0.2),c(0.2),grid_points=100,var.e=1/100)
exlm(x,y,c("EY"),c(-0.2),c(0.2),grid_points=10001,var.e=1/100)
}
