% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exlm_wald.R
\name{exlm_wald.test}
\alias{exlm_wald.test}
\title{Test for general linear hypotheses in the externally informed linear model}
\usage{
exlm_wald.test(
  X,
  y,
  mom,
  I_min,
  I_max = I_min,
  grid_points = 10001,
  intercept = TRUE,
  var.e,
  R,
  r
)
}
\arguments{
\item{X}{Numeric matrix containing independent variables in colums.}

\item{y}{Numeric vector containing dependent variable.}

\item{mom}{Character vector of moments, "EY" = expected value of y, "EXi" = expected value of ith column of X,
"EXiY" = mixed moment of y and x_i, "EY2" = second moment of y, "VY" = variance of y,
"CXiY" = covariance of y and x_i, "RXiY" = correlation of y and x_i, "BXiY" = simple linear regression slope
of y on x_i. The index i must be replaced by a column number of X, e.g. "EX1Y" takes the first column of X.}

\item{I_min}{Numeric vector containing the lower bounds of external interval.}

\item{I_max}{Numeric vector containing the upper bounds of external interval.}

\item{grid_points}{Number of Grid points.}

\item{intercept}{Logical, indicates wheter intercept is used in model or not.}

\item{var.e}{Numeric matrix equal to external variance.}

\item{R}{Numeric matrix defining contrasts R for the nullhypothesis Rb=r.}

\item{r}{Numeric vector r for the nullhypothesis Rb=r.}
}
\value{
list containing a string describing the test result and a p-value.
}
\description{
Tests the general Null hypothesis: R \%*\% beta_0 = r for a full rank matrix R and vector r.
}
\examples{

#overall F-test for simple linear model under external interval for EY:

exlm_wald.test(rnorm(100),rnorm(100),c("EY"),c(-0.2),c(0.2),
grid_points=100,var.e=1/100,R=diag(2),r=c(0,0))

#Simulation:
erg<-c()
for(i in 1:500){erg<-c(erg,exlm_wald.test(rnorm(100),rnorm(100),c("EY"),c(-0.2),c(0.2),
grid_points=100,var.e=1/100,R=diag(2),r=c(0,0))$p.value)}
hist(erg)
}
