% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exSH.R
\name{exSH}
\alias{exSH}
\title{Externally informed Sargan-Hansen test for linear external moment functions}
\usage{
exSH(H, e_min, e_max, var.e, Omega = "Sigma", asymptotic = T)
}
\arguments{
\item{H}{Numeric matrix containing external moment function values, rows are units, columns different moments.}

\item{e_min}{Numeric vector containing the lower bounds of external interval.}

\item{e_max}{Numeric vector containing the upper bounds of external interval.}

\item{var.e}{Numeric matrix equal to external variance.}

\item{Omega}{Character specifying estimator of Omega_h, "S" = sample variance matrix, "Sigma" = obvious estimator.}

\item{asymptotic}{Logical, if true asymptotic distribution is used, if false small sample test is used.}
}
\value{
list containing a string describing the test result and a p-value.
}
\description{
Externally informed Sargan-Hansen test for linear external moment functions
}
\examples{
exSH(rnorm(100),c(-0.4),c(-0.3),1/30)

exSH(cbind(rnorm(100),rnorm(100)),c(-0.2,-0.4),c(0.1,0.3),
Omega="S",asymptotic=FALSE,var.e=1/30*diag(2))
}
