% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exlm_conf.R
\name{interval_gmm}
\alias{interval_gmm}
\title{Confidence Union for a linear model}
\usage{
interval_gmm(
  X,
  y,
  mom,
  I_min,
  I_max = I_min,
  grid_points = 10001,
  intercept = TRUE
)
}
\description{
Confidence Union for a linear model
}
